+++ aa_forum/tests/test_dhooks_lite.py
@@ -7,7 +7,7 @@ from unittest import TestCase
 
 # AA Forum
 from aa_forum import __version__
-from aa_forum.constants import APP_NAME, GITHUB_URL
+from aa_forum.constants import APP_NAME_VERBOSE_USERAGENT, GITHUB_URL
 from aa_forum.helper.discord_messages import _dhooks_lite_user_agent
 
 
@@ -26,7 +26,7 @@ class TestDhooksLiteUserAgent(TestCase):
 
         obj = _dhooks_lite_user_agent()
 
-        self.assertEqual(first=obj.name, second=APP_NAME)
+        self.assertEqual(first=obj.name, second=APP_NAME_VERBOSE_USERAGENT)
         self.assertEqual(first=obj.url, second=GITHUB_URL)
         self.assertEqual(first=obj.version, second=__version__)
 
@@ -41,5 +41,6 @@ class TestDhooksLiteUserAgent(TestCase):
         obj = _dhooks_lite_user_agent()
 
         self.assertEqual(
-            first=str(obj), second=f"{APP_NAME} ({GITHUB_URL}, {__version__})"
+            first=str(obj),
+            second=f"{APP_NAME_VERBOSE_USERAGENT} ({GITHUB_URL}, {__version__})",
         )